Refactor socket messenger base class
