+++ src/TGConnection/SocketMessenger/EncryptedSocketMessenger.php
@@ -13,7 +13,6 @@ use TelegramOSINT\MTSerialization\AnonymousMessage;
 use TelegramOSINT\MTSerialization\MTDeserializer;
 use TelegramOSINT\MTSerialization\OwnImplementation\OwnDeserializer;
 use TelegramOSINT\TGConnection\DataCentre;
-use TelegramOSINT\TGConnection\Socket\PersistentSocket;
 use TelegramOSINT\TGConnection\Socket\Socket;
 use TelegramOSINT\TGConnection\SocketMessenger\EncryptedSocketCallbacks\CallbackMessageListener;
 use TelegramOSINT\TGConnection\SocketMessenger\EncryptedSocketCallbacks\ExpectingMessageListener;
@@ -34,17 +33,8 @@ use TelegramOSINT\TLMessage\TLMessage\TLClientMessage;
 /**
  * Encrypted Telegram-specific socket
  */
-class EncryptedSocketMessenger implements SocketMessenger
+class EncryptedSocketMessenger extends TgSocketMessenger implements SocketMessenger
 {
-    private const HEADER_LENGTH_BYTES = 4;
-    /**
-     * @var Socket
-     */
-    private $socket;
-    /**
-     * @var PersistentSocket
-     */
-    private $persistentSocket;
     /**
      * @var int
      */
@@ -105,8 +95,6 @@ class EncryptedSocketMessenger implements SocketMessenger
      * @var AnonymousMessage[]
      */
     private $messagesToBeProcessedQueue = [];
-    /** @var ReadState|null */
-    private $readState;
 
     /**
      * @param Socket          $socket
@@ -115,8 +103,7 @@ class EncryptedSocketMessenger implements SocketMessenger
      */
     public function __construct(Socket $socket, AuthKey $authKey, MessageListener $callback)
     {
-        $this->socket = $socket;
-        $this->persistentSocket = new PersistentSocket($this->socket);
+        parent::__construct($socket);
         $this->messageReceiptCallback = $callback;
 
         $this->msg_seqno = 0;
@@ -153,45 +140,13 @@ class EncryptedSocketMessenger implements SocketMessenger
 
     /**
      * @throws TGException
-     * @noinspection DuplicatedCode
      */
     protected function readMessageFromSocket(): ?AnonymousMessage
     {
-        if (!$this->readState) {
-            $this->readState = new ReadState();
-        }
-        if (!$this->readState->getLength()) {
-            // header
-            $lengthValue = $this->socket->readBinary(self::HEADER_LENGTH_BYTES);
-            if ($lengthValue === false)
-                return null;
-            if (strlen($lengthValue) != self::HEADER_LENGTH_BYTES)
-                throw new TGException(TGException::ERR_DESERIALIZER_BROKEN_BINARY_READ, self::HEADER_LENGTH_BYTES.'!='.strlen($lengthValue));
-            // data
-            $payloadLength = unpack('I', $lengthValue)[1] - self::HEADER_LENGTH_BYTES;
-            $this->readState->setLengthValue($lengthValue);
-            $this->readState->setLength($payloadLength);
-        } else {
-            $payloadLength = $this->readState->getLength();
-            $lengthValue = $this->readState->getLengthValue();
-        }
-        $lengthToRead = $payloadLength - $this->readState->getCurrentLength();
-        $newPayload = $this->persistentSocket->readBinary($lengthToRead);
-        if (strlen($newPayload)) {
-            $this->readState->addRead($newPayload);
-        }
-        if (!$this->readState->ready()) {
-            $timeDiff = 1000.0 * (microtime(true) - $this->readState->getTimeStart());
-            if ($timeDiff > LibConfig::CONN_SOCKET_TIMEOUT_PERSISTENT_READ_MS) {
-                throw new TGException(TGException::ERR_CONNECTION_SOCKET_READ_TIMEOUT);
-            }
-
+        $packet = $this->readPacket();
+        if (!$packet) {
             return null;
         }
-
-        // full TL packet
-        $packet = $lengthValue.$this->readState->getPayload();
-        $this->readState = null;
 
         return $this->deserializePayload(
             $this->decodeDecryptedPayloadHeaders(
